another commit

+++ models.py
@@ -1,12 +1,13 @@
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime, timezone
 
+
 db = SQLAlchemy()
 
 class User(db.Model):
     __tablename__ = 'users'
     id = db.Column(db.Integer, primary_key=True)
-    chat_id = db.Column(db.Integer, unique=True, nullable=False)
+    chat_id = db.Column(db.BigInteger, unique=True, nullable=False)
     level = db.Column(db.String(80), default='starter', nullable=False)
     levelpoint = db.Column(db.Integer, default=0)
     exchange = db.Column(db.String(180), default='Binance')
@@ -22,8 +23,8 @@ class User(db.Model):
 class Referral(db.Model):
     __tablename__ = 'referrals'
     id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    referred_chat_id = db.Column(db.Integer, nullable=False)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('users.id'), nullable=False)
+    referred_chat_id = db.Column(db.BigInteger, nullable=False)
     timestamp = db.Column(db.DateTime, default=datetime.now(timezone.utc))
 
 class Blockchains(db.Model):
